+++ charts-project/tennis.h
@@ -2,15 +2,13 @@
 #define TENNIS_H
 
 #include "repetition.h"
-
 class Tennis : public Repetition
 {
 public:
     Tennis(const QDateTime& = QDateTime(QDate(2000, 1, 1), QTime(0,0)), double = minWeight);
-    unsigned int Intensity() const override;
+    double Intensity() const override;
     unsigned int CaloriesBurned() const override;
     Tennis* clone() const override;
-
 };
 
 #endif // TENNIS_H
